refactor(FIX-004): mover domain/vo/ para domain/enums/ — eliminar pacote duplicado
CategoriaPedido e FormaPagamento migrados de domain/vo/ para domain/enums/,
unificando a convencao de pacotes do dominio. Pasta domain/vo/ deletada.
18 arquivos com imports atualizados via sed em lote.
mvn compile + 321 testes unitarios verdes, zero regressoes.
Item de PENDENCIAS-TECNICAS.md marcado como resolvido.
