+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/application/dto/FuncionarioResponse.java
@@ -1,7 +1,7 @@
 package br.com.satyan.stering.saita.financasbottelegram.application.dto;
 
 import br.com.satyan.stering.saita.financasbottelegram.domain.entity.Funcionario;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.FormaPagamento;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.FormaPagamento;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
