+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/application/dto/ValeResponse.java
@@ -2,7 +2,7 @@ package br.com.satyan.stering.saita.financasbottelegram.application.dto;
 
 import br.com.satyan.stering.saita.financasbottelegram.domain.enums.StatusPedido;
 import br.com.satyan.stering.saita.financasbottelegram.domain.model.PedidoPagamento;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.CategoriaPedido;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.CategoriaPedido;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
