+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/application/services/CadastrarFuncionarioServiceImpl.java
@@ -3,7 +3,7 @@ package br.com.satyan.stering.saita.financasbottelegram.application.services;
 import br.com.satyan.stering.saita.financasbottelegram.application.port.in.CadastrarFuncionarioPortIn;
 import br.com.satyan.stering.saita.financasbottelegram.application.port.out.FuncionarioRepositoryPortOut;
 import br.com.satyan.stering.saita.financasbottelegram.domain.entity.Funcionario;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.FormaPagamento;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.FormaPagamento;
 import org.springframework.stereotype.Service;
 
 /**
