+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/application/services/CadastrarValeServiceImpl.java
@@ -6,7 +6,7 @@ import br.com.satyan.stering.saita.financasbottelegram.application.port.out.Pedi
 import br.com.satyan.stering.saita.financasbottelegram.domain.enums.StatusPedido;
 import br.com.satyan.stering.saita.financasbottelegram.domain.exceptions.FuncionarioNaoEncontradoException;
 import br.com.satyan.stering.saita.financasbottelegram.domain.model.PedidoPagamento;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.CategoriaPedido;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.CategoriaPedido;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import org.springframework.stereotype.Service;
