+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/application/services/FecharMesServiceImpl.java
@@ -10,7 +10,7 @@ import br.com.satyan.stering.saita.financasbottelegram.domain.enums.StatusPedido
 import br.com.satyan.stering.saita.financasbottelegram.domain.exceptions.FechamentoDuplicadoException;
 import br.com.satyan.stering.saita.financasbottelegram.domain.exceptions.FuncionarioNaoEncontradoException;
 import br.com.satyan.stering.saita.financasbottelegram.domain.model.PedidoPagamento;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.CategoriaPedido;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.CategoriaPedido;
 import java.math.BigDecimal;
 import java.text.NumberFormat;
 import java.time.LocalDate;
