+++ financas_bot_telegram/src/main/java/br/com/satyan/stering/saita/financasbottelegram/domain/entity/Funcionario.java
@@ -1,6 +1,6 @@
 package br.com.satyan.stering.saita.financasbottelegram.domain.entity;
 
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.FormaPagamento;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.FormaPagamento;
 import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.AllArgsConstructor;
