+++ financas_bot_telegram/src/test/java/br/com/satyan/stering/saita/financasbottelegram/adapters/in/rest/funcionario/FuncionarioControllerTest.java
@@ -12,7 +12,7 @@ import br.com.satyan.stering.saita.financasbottelegram.application.port.in.Cadas
 import br.com.satyan.stering.saita.financasbottelegram.application.port.out.FuncionarioRepositoryPortOut;
 import br.com.satyan.stering.saita.financasbottelegram.domain.entity.Funcionario;
 import br.com.satyan.stering.saita.financasbottelegram.domain.exceptions.FuncionarioNaoEncontradoException;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.FormaPagamento;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.FormaPagamento;
 import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
