+++ financas_bot_telegram/src/test/java/br/com/satyan/stering/saita/financasbottelegram/application/services/CadastrarAdiantamentoServiceImplTest.java
@@ -11,7 +11,7 @@ import br.com.satyan.stering.saita.financasbottelegram.application.port.out.Func
 import br.com.satyan.stering.saita.financasbottelegram.domain.entity.Adiantamento;
 import br.com.satyan.stering.saita.financasbottelegram.domain.entity.Funcionario;
 import br.com.satyan.stering.saita.financasbottelegram.domain.exceptions.FuncionarioNaoEncontradoException;
-import br.com.satyan.stering.saita.financasbottelegram.domain.vo.FormaPagamento;
+import br.com.satyan.stering.saita.financasbottelegram.domain.enums.FormaPagamento;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Optional;
